+++ src/components/weather/Weather.tsx
@@ -1,25 +1,12 @@
-import { useEffect, useState } from 'react';
 import WeatherDisplay from './WeatherDisplay';
+import useWeather from '@/hooks/useWeather';
 
 const Weather = () => {
-  const [weather, setWeather] = useState();
-  useEffect(() => {
-    getWeather();
-  }, []);
-
-  async function getWeather() {
-    const data = await fetch(
-      'https://api.open-meteo.com/v1/forecast?latitude=28&longitude=84&current=temperature_2m,relative_humidity_2m,rain,wind_direction_10m&format=json&timeformat=unixtime',
-    );
-
-    const res = await data.json();
-    setWeather(res.current);
-  }
-
+  const { weather, isLoading } = useWeather(84, 28);
   return (
     <>
       <div className="flex h-full items-center justify-center bg-neutral-300">
-        {weather ? <WeatherDisplay weather={weather} /> : 'Loading...'}
+        {isLoading ? 'Loading...' : <WeatherDisplay weather={weather} />}
       </div>
     </>
   );
